fix sort
