+++ frontend/src/reducers/post.js
@@ -11,10 +11,6 @@ import {
   EDIT_COMMENT_SUCCESS,
   UPVOTE_COMMENT_SUCCESS,
   DOWNVOTE_COMMENT_SUCCESS,
-  SORT_BY_DATE_DESC,
-  SORT_BY_VOTE_DESC,
-  SORT_BY_DATE_ASC,
-  SORT_BY_VOTE_ASC,
 } from './../actions/constants';
 import comments from './comment';
 
@@ -22,14 +18,6 @@ const posts = (state = [], action) => {
   switch (action.type) {
     case GET_ALL_POSTS_SUCCESS:
       return action.posts;
-    case SORT_BY_DATE_ASC:
-      return [...state.sort(compareTime)];
-    case SORT_BY_VOTE_ASC:
-      return [...state.sort(compareVoteScore)];
-    case SORT_BY_DATE_DESC:
-      return [...state.sort(compareTimeDesc)];
-    case SORT_BY_VOTE_DESC:
-      return [...state.sort(compareVoteScoreDesc)];
     case CREATE_POST_SUCCESS:
       return [
         ...state,
@@ -122,28 +110,5 @@ const post = (state = {}, action) => {
   }
 };
 
-function compareTime(a, b) {
-  if (a.timestamp < b.timestamp) { return -1; }
-  if (a.timestamp > b.timestamp) { return 1; }
-  return 0;
-}
-
-function compareVoteScore(a, b) {
-  if (a.voteScore < b.voteScore) { return -1; }
-  if (a.voteScore > b.voteScore) { return 1; }
-  return 0;
-}
-
-function compareTimeDesc(a, b) {
-  if (a.timestamp < b.timestamp) { return 1; }
-  if (a.timestamp > b.timestamp) { return -1; }
-  return 0;
-}
-
-function compareVoteScoreDesc(a, b) {
-  if (a.voteScore < b.voteScore) { return 1; }
-  if (a.voteScore > b.voteScore) { return -1; }
-  return 0;
-}
 
 export default posts;
